Issue/62 (#64)
Parameter checking to see if we should update a provisioned product did not consider the default values of parameters in the CloudFormation template.  These default values are used when no others are specified.  This was causing Service Catalog to run a product plan that thought there were changes but CloudFormation was aware there were no changes.  The CloudFormation changeset was invalid and caused ServiceCatalog to taint the product. issues #63 #62